Remove references to print out and print tests, add method to compare objects
Remove test to check if printPosition works, add method to check if position objects are equal, change tests for end coordinates of various rovers to use new method to check if two positions are the same and pass tests.  All in the vein of fixing problem with test runner not being able to run RoverTest because of streaming issues.